rename internal/licensing to internal/accesstoken
