+++ enterprise/cmd/cody-gateway/internal/actor/productsubscription/BUILD.bazel
@@ -11,8 +11,8 @@ go_library(
         "//enterprise/cmd/cody-gateway/internal/dotcom",
         "//enterprise/internal/licensing",
         "//enterprise/internal/productsubscription",
+        "//internal/accesstoken:licensing",
         "//internal/codygateway",
-        "//internal/licensing",
         "//internal/trace",
         "//lib/errors",
         "@com_github_gregjones_httpcache//:httpcache",
