+++ enterprise/cmd/frontend/internal/dotcom/productsubscription/BUILD.bazel
@@ -29,6 +29,7 @@ go_library(
         "//enterprise/internal/license",
         "//enterprise/internal/licensing",
         "//enterprise/internal/productsubscription",
+        "//internal/accesstoken:licensing",
         "//internal/auth",
         "//internal/authz",
         "//internal/codygateway",
@@ -42,7 +43,6 @@ go_library(
         "//internal/featureflag",
         "//internal/gqlutil",
         "//internal/hashutil",
-        "//internal/licensing",
         "//internal/redispool",
         "//internal/slack",
         "//internal/types",
@@ -88,6 +88,7 @@ go_test(
         "//enterprise/internal/license",
         "//enterprise/internal/licensing",
         "//enterprise/internal/productsubscription",
+        "//internal/accesstoken:licensing",
         "//internal/actor",
         "//internal/auth",
         "//internal/authz",
@@ -97,7 +98,6 @@ go_test(
         "//internal/errcode",
         "//internal/featureflag",
         "//internal/hashutil",
-        "//internal/licensing",
         "//internal/redispool",
         "//internal/slack",
         "//internal/timeutil",
