+++ enterprise/internal/licensing/BUILD.bazel
@@ -20,13 +20,13 @@ go_library(
     visibility = ["//enterprise:__subpackages__"],
     deps = [
         "//enterprise/internal/license",
+        "//internal/accesstoken:licensing",
         "//internal/conf",
         "//internal/conf/conftypes",
         "//internal/env",
         "//internal/errcode",
         "//internal/goroutine",
         "//internal/httpcli",
-        "//internal/licensing",
         "//internal/redispool",
         "//lib/errors",
         "@com_github_derision_test_glock//:glock",
