+++ internal/accesstoken/BUILD.bazel
@@ -4,7 +4,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "licensing",
     srcs = ["access_token.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/internal/licensing",
+    importpath = "github.com/sourcegraph/sourcegraph/internal/accesstoken",
     visibility = ["//:__subpackages__"],
     deps = [
         "//internal/hashutil",
@@ -13,9 +13,9 @@ go_library(
 )
 
 go_test(
-    name = "licensing_test",
+    name = "accesstoken_test",
     srcs = ["access_token_test.go"],
-    embed = [":licensing"],
+    embed = [":accesstoken"],
     deps = [
         "@com_github_hexops_autogold_v2//:autogold",
         "@com_github_stretchr_testify//assert",
